Fix cilium install when using get_kubeconfig.py
At first, it was believed that the kubeconfig created with
`get_kubeconfig.py` was bugged due to permission/scoping issues
with the bearer token. After further discussion and investigation,
it was determined that the cilium-cli was not autodetecting the
kubernetes installation to be on gke. After modifications, the
`pipeline.yml` task can now be executed successfully.

cilium-cli determines that the target cluster is gke based on
the cluster name within the kubeconfig (see
https://github.com/cilium/cilium-cli/blob/master/k8s/client.go#L495).
cilium-cli determines the CIDR used for native routing using gcloud,
which is why it is set manually (see
https://github.com/cilium/cilium-cli/blob/master/install/gke.go#L19).

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>