Create toolkit go package
The toolkit go package aims to provide convient functionality for
common tasks related to running performance benchmarks. The
functionality added to toolkit should be something that is not
convient or do-able from bash, as toolkit is just meant to
automate tasks which would be difficult to do outside of a
script.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>