Add kustomize templates for pod2pod network tests
kustomize/topologies is meant to store kustomization files that
allow for quickly creating pod topologies for network-based
micro-benchmarks. pod2pod/base provides a simple deployment
which can be edited by overlays to configure layout of
pods.

These templates should be modified as needed for performing
benchmarks, they are not indented to be used directly.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>