Add functionality and command for fetching crane
crane is a useful binary for working with container registries,
developed by google and hosted over at
github.com/google/go-containerregistry. With the crane bin
easily accessible, more functionality can be added to
scaffolding for creating an in-cluster registry which crane can
push images to. This will enable a faster development process,
as: (1) crane has support for pushing image tarballs to
registries, simplifying the process for using images built
with nix's docker tools, (2) upload and download bandwidth
and time does not have to spent pushing and pulling images
from quay.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>