Modify base pod2pod topology to use deployments
By using deployments instead of individual pods, we can scale
the number of server/client pods for different kinds of tests.
For instance if we want to build a pod-service test, we can
just increment the number of replicas on the server deployment
and create a server that targets them.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>