Remove cilium & crane lazyget from toolkit
The ability to download cilium-cli and crane binaries was
originally planned to be placed into the toolkit, however in
hindsight this functionality can easily be added into scaffolding
as some bash scripts. By making this migration, the amount of go
code that must be maintained and tested is reduced, saving time
and complexity in the long run.

Scaffolding works on the premise that the less 'code' the better,
prefering scripts which reflect how a user would do something
in their shell. This commit is an instance where this philosophy
took precedence.

If it is determined this functionality needs to be added back
into the toolkit, the code can be picked out of the git history.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>