Add handling for cafile in k8s-api-readyz.sh
k8s-api-readyz.sh only supported handling cases where the ca
data was stored in the kubeconfig, rather than on file. This
commit adds support for reading configured cafiles (ie
"certificate-authority" key in the kubeconfig), for when
the cafile data (ie "certificate-authority-data") key is not set.
"certificate-authority-data" takes precedence. If neither key is
set, then '--insecure' is passed to curl to still allow the check
to take place.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>