Add simple template for pod2pod service scenario
Adds an overlay with a simple service definition. This is
definitely subject to change in the future, as users still have
to fill out the service with the port they want to target and
the service type. The template right now just provides some
structure and a simple starting point.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>