Switch to lb svc and add cmds for *.netfilter=0
When using netperf it is important to know that the port the
client uses must match the port the server is listening on. If
the client tries to connect to port 30000, say a node port, but
the server is only listening on 12865, then the server will
reject the connection. Using a loadbalancer gets around this
issue while also making for a more realistic test.

Additionally, commands were added to ensure that netfilter
is set to zero in sysctl, to rule out said issue when trying
to debug network issues.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>