Add metallb for xdp test, fix netperf 0 t/s
Rather than rely on minikube tunnel, use metallb. Not exactly sure
how we can measure the impact of xdp acceleration with dsr on cpu,
but TDB.

Additionally fixes an issue with the netperf test, as I was
getting a consistent 0 transactions / second. TIL netperf uses port
12865 to establish a connection, then uses another port to
transfer data. The k8s service was only setup to allow for traffic
coming into the 'connection control' port. I added another
explicit port in the service, 12866, for data transfer.

Another aspect to this commit is adding in the '-k' arg to netperf
for printing additional stats. This was taken from
cilium/kubenetbench, thank you!

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>