Clean up inconsistencies in scripts/
A lot of these scripts use `-d` as a debug parameter, but there
is a floating `VERBOSE=""` around this code, left in from a bad
copy and paste from `k8s_api_readyz.sh`. Also, added a
top level comment to `retry.sh` and renamed `k8s-api-readyz.sh`
to match the underscore convention of other scripts.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>