Upgrade grafana dashboards (#56)
* Use patches for adding dashboards over configmaps

Individual configmaps were being used to add new dashboards into
grafana, requiring adding new mounts into the grafana pod. This
change introduces using a single configmap to store dashboards,
allowing for kustomize patches to be used to modify the configmap
and therefore the dashboards passed into grafana in a more
streamlined way.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Remove docker monitoring dashboards

Too outdated to be working with cadvisor, no metrics show. We need
something different.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Remove node-exporter dashboard json

This text is already stored in the included configmap, no reason
it should also exist in its own file here. If such a reason comes
up though, this change can be reverted.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Add dashboards for node-exporter and cilium v12

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Add script to include new grafana dashboards

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Mark dashboards as writable for editing

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Capitalize Prometheus datasource name

A lot of dashboards use a default prometheus datasource name of
"Prometheus" over "prometheus". This commit does the
capitalization to make it easier to import dashboards in the
future, without having to go through the hassle of adjusting the
datasource if necessary.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Move node-exporter and cadvisor out of prom dir

It was once helpful to have these resources nested together to
let one kustomization file in the root apply all of them. This
has made it more difficult to test out changes to individual
components, as they were biased to be applied all-together
rather than individually.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

* Update grafana README

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>