k8s/helper: Modify WaitOnWatchedResource to take ListOptions struct
Before this commit, Helper.WaitOnWatchedResource took the name
and namespace of a resource to watch, constructing a ListOptions
struct internally to match on `metadata.name`. This made watching
multiple types of resources or using custom selectors impossible.
This commit replaces the name parameter with a listOptions one,
allow for a user to provide a custom ListOptions struct.

This commit also adds a new function, NewListOptionsFromName, to
construct a ListOptions struct that targets a given name through
a field selector on `metadata.name`.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>