k8s/helper: Modify DeleteResourceAndWaitGone to use given name and ns
`Helper.DeleteResourceAndWaitGone` used to require users to have a
copy of the unstructured resource they wish to delete, which may
not always be available. This commit modifies the function to let
users pass the name and namespace of the resource they want to delete
manually, without needing a copy of the resource.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>