cmapisrv-mock: initial implementation
Implement the logic to mock the behavior of the clustermesh-apiserver.
It supports mocking a user-configurable number of clusters, and
leverages the kvstoremesh API to allow storing the information of
multiple clusters using a single etcd instance. Each cluster is composed
of a configurable number of nodes, endpoints, identities and services,
which are initialized at start-up time. Subsequently, at run-time, it
additionally generates insertion, update (where appropriate) and
deletion churn, depending on the configured QPS values.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>