toolkit: Add option to verify k8s-ready to ignore nodes in a cluster
This commit adds the --ignored-nodes option to the verify k8s-ready
subcommand of the toolkit. This argument can be used to ignore the
status of certain nodes in a cluster. An example use case is a
deployment where Cilium is purposefully not scheduled onto a node,
causing said node to never enter into a ready state.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>