cmapisrv-mock: fix incorrect nodes and services path
Currently, the mocked nodes and services prefixes incorrectly contain
the cluster name twice, as both explicitly configured as part of the
sync store prefix and automatically added as part of the key generation.
This mistake got previously unnoticed, as the base prefix was correct
and, historically, Cilium did not use the key during unmarshalling.
However, the introduction of more thorough validation brought this
issue to light, and is now being fixed.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>