components: Add egw-scale-utils for scale testing EGW
This commit adds a new component for scale testing Cilium's Egress
Gateway feature. This initial commit introduces the masquerade delay
test, which determines the amount of time it takes for traffic egressing
from a new workload Pod to be masqueraded through an EGW node. The
details of the test and its implementation can be found in the included
README.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>