ci: Adjust logic for determining build tag
This commit adjusts the logic used to determine the image tag in the
build workflow. Previously, only the SHA of built ref was used,
requiring image tags in manifests to be manually updated. This commit
will use the 'latest' tag instead if the ref being built is on the main
branch.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>