egw-scale-utils: ensure consistent interval, and don't return on errors
Modify the client loop to ensure that probes are sent evenly spread,
with the desired interval (which is now configurable as a parameter),
independently of the time spent establishing the connection or reading
the data from it. Previously, instead, the fixed 50ms sleep could lead
to skews, making results dependent on the actual network latency.
Similarly, reading from the connection could block forever, as no
deadline was explicitly set.

Additionally, let's change all connection errors to not be fatal, so
that the client doesn't restart in case of timeouts. We were indeed
observing this behavior in the actual tests (likely caused by SYN
packets being dropped during datapath reconfiguration), partially
invalidating the results. Errors are now instead treated as misses.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>