egw-scale-utils/client: introduce stress mode
Extend the egw-scale-utils client with a new mode that allows to assess
the maximum number of parallel connections that can be opened from a
client to an external destination via the egress gateway. Specifically,
once the first connection is successfully established, the client
continuously opens new connections until repeated failures are
encountered. All connections are closed only once the client process
gets terminated.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>