egw-scale-utils/server: add option to keep connection open
Extend the egw-scale-utils server with the possibility of keeping client
connections open, rather then closing them immediately as soon as the
reply is sent back. This mode is useful to assess the maximum number of
parallel connections that can be opened from a client to an external
destination via the egress gateway.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>