egw-scale-utils: drop test manifests
The CL2 configuration has been moved to the cilium/cilium repository,
alongside the workflow definition. Let's remove the copy still available
here to reduce the maintenance burden, considering that it is already
out of sync. The README file is also simplified, as the full description
of the environment is available in the cilium/cilium repository.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>