cmapisrv-mock: fix handling of the bootstrap process
The blamed commit bumped the Cilium dependency to v1.18. However, it
missed to update the handling around the bootstrap process, causing
the mocker container to turn ready immediately, instead of waiting
until the initial state has been populated to etcd, and consequently
also lowering the rate limiting settings to the runtime ones. Let's
get this fixed to restore the previous behavior.

Fixes: 6973fa4b ("fix(deps): update all go dependencies")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>